consumption creation implementation
